requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
groq>=0.9.0
python-dotenv>=1.0.0
